// types/repository.d.ts
export interface Repository {
    // name: string;
    // description?: string;
    // url: string;
    // html_url: string;
    // stargazers_count: number;
    // forks_count: number;
    // watchers_count: number;
    // open_issues_count: number;
    // language?: string;
    // pushed_at?: string;
    name: string;
    description?: string;
    url: string;
    html_url: string;
    stargazers_count: number;
    forks_count: number;
    watchers_count: number;
    language?: string;
    pushed_at: string;
    created_at: string;
    default_branch?: string;
    open_issues_count: number;
}